+++ mllib/src/main/scala/org/apache/spark/ml/classification/GBTClassifier.scala
@@ -257,7 +257,7 @@ final class GBTClassificationModel private[ml](
   }
 }
 
-private[spark] object GBTClassificationModel {
+private[ml] object GBTClassificationModel {
 
   /** (private[ml]) Convert a model from the old API */
   def fromOld(
